[heap] Make FreeList less implementation-dependent
This CL is the first step toward implementing map space's freelist
using a heap rather than a linked list. The next step would be to
implement a class FreeListHeap (inheriting from FreeList).

However, the GC assumes in different places that FreeList::RemovePage
can remove any page from the freelists. While this is true with a
linked list, it's not with a heap. Therefore, this CL will not be
commited.

Furthermore, this proposed implementation is not ideal: some methods
defined in FreeListLinked should probably rather be in FreeList
(Available, IsEmpty...). Ideally, FreeListLinked would just implement
ForAllFreeListCategories, which would then be used by FreeList for
those methods. However, because ForAllFreeListCategories is templated,
it cannot be virtual, which makes things a bit hard. Using templates
can overcome this, but has it's own dowsides.